requirements.txt
PyYAML==3.12
jinja2
boto3
flask
function-shield